Add `PushLineSplitter`.
This implements a strategy for doing platform agnostic line-splitting
in a push-oriented manner. It can be caught out if text happens to
contain CR or LF characters which aren't intended to be a newline
(which seems pretty dubious anyway), but should never result in
the addition or removal of newlines otherwise.