Add a special tag to path patterns when overriding the workspace.
If a root path is specified, we need to override both the current
working directory and `$GITHUB_WORKSPACE` during cache
save, peek and restore. By incoporating a special comment into the list
of path patterns, we can also reflect this in the path hash and avoid
conflicts with cache entries where this was not done.